Move generatepdf logic to job
