+++ app/Jobs/GenerateDetailReport.php
@@ -1,47 +1,51 @@
 <?php
 
-namespace App\Livewire\Pages\Table;
+namespace App\Jobs;
 
+use Carbon\Carbon;
 use App\Models\Device;
 use App\Models\Report;
-use Livewire\Component;
 use App\Models\Location;
 use App\Models\Measurement;
 use App\Models\TrainProfile;
-use App\Jobs\GenerateDetailReport;
+use Illuminate\Bus\Queueable;
 use Spatie\LaravelPdf\Facades\Pdf;
 use Spatie\Browsershot\Browsershot;
 use Spatie\LaravelPdf\Enums\Format;
 use Illuminate\Support\Facades\File;
+use Illuminate\Queue\SerializesModels;
+use Illuminate\Queue\InteractsWithQueue;
+use Illuminate\Contracts\Queue\ShouldQueue;
+use Illuminate\Foundation\Bus\Dispatchable;
 
-class TableDetail extends Component
+class GenerateDetailReport implements ShouldQueue
 {
+    use Dispatchable, InteractsWithQueue, Queueable, SerializesModels;
 
-    public $id;
+    protected $id;
+    protected $reportId;
 
-    public function mount($id)
+    /**
+     * Create a new job instance.
+     */
+    public function __construct($id, $reportId)
     {
-        $this->id($id);
+        $this->id = $id;
+        $this->reportId = $reportId;
     }
 
-    public function render()
-    {
-        $train = TrainProfile::with('stations')->where('device_id', $this->id)->first();
-        $device = Device::selectRaw('*, ST_X(last_location::geometry) AS longitude, ST_Y(last_location::geometry) AS latitude')->where('serial_number', $this->id)->first();
-        $device->last_monitored_value = json_decode($device->last_monitored_value, true);
-        $locations = Location::selectRaw('*, ST_X(point::geometry) AS longitude, ST_Y(point::geometry) AS latitude')->where('device_id', $this->id)->orderBy('created_at', 'desc')->paginate(5, pageName: 'locationsPage');
-        $measurements = Measurement::where('device_id', $this->id)->orderBy('created_at', 'desc')->paginate(5, pageName: 'measurementsPage');
-        return view('livewire.pages.table.table-detail', ['train' => $train, 'device' => $device, 'locations' => $locations, 'measurements' => $measurements]);
-    }
-
-    public function generatePdf()
+    /**
+     * Execute the job.
+     */
+    public function handle(): void
     {
         $reportsDirectory = base_path('public/reports');
+        $now = Carbon::now()->format('Y-m-d(h_i_s)');
 
         if (!File::exists($reportsDirectory)) {
             File::makeDirectory($reportsDirectory, 0755, true);
         }
-        $pdfFilePath = $reportsDirectory . "/" . "Test" . ".pdf";
+        $pdfFilePath = $reportsDirectory . "/" . "Device_" . $this->id . "_" . $now . ".pdf";
 
         $train = TrainProfile::with('stations')->where('device_id', $this->id)->first();
         $device = Device::selectRaw('*, ST_X(last_location::geometry) AS longitude, ST_Y(last_location::geometry) AS latitude')->where('serial_number', $this->id)->first();
@@ -70,22 +74,15 @@ class TableDetail extends Component
             ->format(Format::A4)
             ->landscape()
             ->save($pdfFilePath);
-    }
 
-    public function createReport()
-    {
-        $reportId = $this->saveReport();
-        GenerateDetailReport::dispatch($this->id, $reportId);
-    }
-
-    public function saveReport()
-    {
-        $report = new Report([
-            "name" => "Generating name",
-            "file" => "Generating report"
-        ]);
-        $report->save();
-
-        return $report->id;
+        $report = Report::find($this->reportId);
+        if ($report) {
+            $report->file = "reports/Device_" . $this->id . "_" . $now . ".pdf";
+            $report->name = "Device_" . $this->id . "_" . $now . ".pdf";
+            $report->save();
+            // ReportingChannel::dispatch($this->authId, 'generating report finished', false);
+        } else {
+            throw new \Exception("Report with ID {$this->reportId} not found");
+        }
     }
 }
